Let each test helper hold one thing
Where a tile is, how tiles are laid into a map, who is placed in a level,
and the particular stages tests are set on now each have their own helper.

- tile_positions.hpp: TestTileSize (now an int, the size the engine takes)
  and feetOf, topLeftOf, middleOf, surfaceOf
- tiles.hpp: only laying tiles into a map
- levels.hpp: spawns, beats and aLevelPlacing; aVillagerAt is aRatAt, since
  a rat is what it spawns
- floor_level.hpp and ledge_and_wall.hpp: the two shared stages
- navigation_maps.hpp: maps.hpp, with the graph queries that only make
  sense on those maps, so its FloorRow and PlatformRow no longer sit
  beside other files' own
- shipped.hpp: tilesOfLevel, which reads a shipped level
- Two UI tests set on the floor level stop rebuilding it by hand

+++ tests/game/test_world.cpp
@@ -3,7 +3,9 @@
 #include "actor/abilities/swing_ability_data.hpp"
 #include <optional>
 #include "game/level_data.hpp"
-#include "helpers/tiles.hpp"
+#include "helpers/tile_positions.hpp"
+#include "helpers/levels.hpp"
+#include "helpers/floor_level.hpp"
 #include "helpers/actors.hpp"
 #include <cstddef>
 #include <memory>
@@ -33,7 +35,6 @@
 #include "game/score.hpp"
 #include "pickups/pickup_data.hpp"
 #include "pickups/pickup_spawn_data.hpp"
-#include "helpers/levels.hpp"
 #include "helpers/palettes.hpp"
 #include "helpers/temporary_levels.hpp"
 #include "helpers/npc_fixtures.hpp"
@@ -232,7 +233,7 @@ TEST_CASE("A spawn moved in the level data is where the npc stands", "[World]")
     world.loadLevel("levels/level6.json");
 
     glm::vec2 spawnAt = spawnsIn(world.getLevel())[1].feet;
-    glm::vec2 movedTo{spawnAt.x - TestTileSize, spawnAt.y};
+    glm::vec2 movedTo{spawnAt.x - static_cast<float>(TestTileSize), spawnAt.y};
 
     LevelData edited = world.getLevelData();
     edited.npcs[1].feet = movedTo;
